Added media recording / playback triggers based on the current media time. Added support for the updated ogg and wav recording library,

+++ gwt-cordova/src/main/java/nl/mpi/tg/eg/experiment/client/listener/MediaTriggerListener.java
@@ -36,7 +36,7 @@ public class MediaTriggerListener {
         return isFirst;
     }
 
-    public boolean triggerWhenReady(Long currentTime) {
+    public boolean triggerWhenReady(Double currentTime) {
         while (currentTime >= currentKey) {
             listenerMap.remove(currentKey).postLoadTimerFired(null);
             if (listenerMap.isEmpty()) {
